Vision-AI (Model Training)/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0.0
matplotlib>=3.7.0
pillow>=9.5.0
# Optional: needed only for video tracking in detection.py
deep-sort-realtime>=1.3.2
